+++ account/forms.py
@@ -7,7 +7,7 @@ from django.core.validators import (
 )
 from django.utils.translation import gettext_lazy as _
 
-from .models import Account, Profile, Tweet
+from .models import Account, Profile
 
 
 class SignUpForm(UserCreationForm):
@@ -90,24 +90,3 @@ class ProfileForm(forms.ModelForm):
         fields = {"profile"}
         widgets = {"profile": forms.Textarea}
         labels = {"profile": "プロフィール"}
-
-
-class TweetForm(forms.ModelForm):
-    """
-    ツイートフォーム
-    """
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-    class Meta:
-        model = Tweet
-        fields = {"content"}
-        widgets = {
-            "content": forms.Textarea,
-            "description": forms.Textarea(attrs={"rows": 6, "cols": 15}),
-        }
-        labels = {"content": "内容"}
-        error_messages = {
-            "content": {"required": "入力が必須です．", "max_length": "文字数は，255文字以下です．"}
-        }
