+++ account/models.py
@@ -21,17 +21,9 @@ class Profile(models.Model):
     user = models.OneToOneField(
         Account, on_delete=models.CASCADE, related_name="profile"
     )
-
     profile = models.TextField(_("profile"), default="")
+    created_at = models.DateTimeField("作成日時", auto_now_add=True, blank=True)
+    updated_at = models.DateTimeField("更新日時", auto_now=True, blank=True)
 
     def __str__(self):
         return self.profile
-
-
-class Tweet(models.Model):
-    user = models.ForeignKey(Account, on_delete=models.CASCADE, related_name="tweet")
-    content = models.CharField(_("content"), max_length=255)
-    created_at = models.DateTimeField(verbose_name="投稿日時", auto_now_add=True)
-
-    def __str__(self):
-        return self.content
