+++ tweet/templates/tweet/tweet_detail.html
@@ -1,0 +1,20 @@
+{% extends 'account/base.html' %}
+
+{% block content %}
+
+  <div class="card">
+    <div class="card-body">
+      <h5 class="card-title">{{ tweet.user }}</h5>
+      <p class="card-text">{{ tweet.content | linebreaksbr }}</p>
+      <p class="card-text" "fs-6"><font color="silver">{{ tweet.created_at }}</font></p>
+      <div align="right">
+        {% if tweet.user == request.user %}
+          <a type="button" class="btn btn-outline-danger" data-mdb-ripple-color="dark" href="{% url 'tweet:delete_tweet' tweet.pk %}">
+            削除
+          </a>
+        {% endif %}
+      </div>
+    </div>
+  </div>
+
+{% endblock %}
